Turn refresh rate into a seek bar preference

+++ app/src/main/java/com/farmerbb/taskbar/service/TaskbarService.java
@@ -318,7 +318,7 @@ public class TaskbarService extends Service {
             }
         });
 
-        refreshInterval = Integer.parseInt(pref.getString("refresh_frequency", "2")) * 1000;
+        refreshInterval = (int) (Float.parseFloat(pref.getString("refresh_frequency", "2")) * 1000);
         sortOrder = pref.getString("sort_order", "false");
 
         switch(pref.getString("recents_amount", "past_day")) {
